@@ -1,17 +1,14 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router/index'
 
-import ElementUI from 'element-ui'
-import 'element-ui/lib/theme-chalk/index.css'
-Vue.use(ElementUI)
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css'
 
 import SkuForm from './components/index'
-Vue.use(SkuForm)
 
-Vue.config.productionTip = false
-
-new Vue({
-    router,
-    render: h => h(App)
-}).$mount('#app')
+const app = createApp(App)
+app.use(router)
+app.use(ElementPlus)
+app.use(SkuForm)
+app.mount('#app')
