@@ -1,10 +1,11 @@
 <template>
+    <!-- eslint-disable vue/multi-word-component-names -->
     <div>
         <SkuForm
+            v-model:attribute="attribute"
+            v-model:sku="sku"
             :source-attribute="sourceAttribute"
             :structure="structure"
-            :attribute.sync="attribute"
-            :sku.sync="sku"
         />
         <el-row type="flex" :gutter="20">
             <el-col>
@@ -20,6 +21,7 @@
 </template>
 
 <script>
+/* eslint-disable vue/multi-word-component-names */
 export default {
     data() {
         return {
